refactor: auth 모듈 admin 모듈에 병합
