+++ admin/src/main/java/kernel360/ckt/admin/application/service/command/ReissueCommand.java
@@ -1,4 +1,4 @@
-package kernel360.ckt.auth.application.service.command;
+package kernel360.ckt.admin.application.service.command;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
