+++ admin/src/main/java/kernel360/ckt/admin/infra/jpa/RefreshTokenJpaRepository.java
@@ -1,8 +1,8 @@
-package kernel360.ckt.auth.infra;
+package kernel360.ckt.admin.infra.jpa;
 
+import java.util.Optional;
 import kernel360.ckt.core.domain.entity.RefreshTokenEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.Optional;
 
 public interface RefreshTokenJpaRepository extends JpaRepository<RefreshTokenEntity, Long> {
     Optional<RefreshTokenEntity> findByToken(String token);
