+++ admin/src/main/java/kernel360/ckt/admin/ui/AuthController.java
@@ -1,13 +1,17 @@
-package kernel360.ckt.auth.ui;
+package kernel360.ckt.admin.ui;
 
 import jakarta.validation.Valid;
-import kernel360.ckt.auth.application.service.AuthService;
-import kernel360.ckt.auth.ui.dto.request.LoginRequest;
-import kernel360.ckt.auth.ui.dto.request.ReissueRequest;
-import kernel360.ckt.auth.ui.dto.response.TokenResponse;
+import kernel360.ckt.admin.application.service.AuthService;
+import kernel360.ckt.admin.ui.dto.request.LoginRequest;
+import kernel360.ckt.admin.ui.dto.request.ReissueRequest;
+import kernel360.ckt.admin.ui.dto.response.TokenResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/api/v1/auth")
