+++ admin/src/main/java/kernel360/ckt/admin/ui/dto/response/TokenResponse.java
@@ -1,0 +1,6 @@
+package kernel360.ckt.admin.ui.dto.response;
+
+public record TokenResponse(
+    String accessToken,
+    String refreshToken
+) {}
