--- auth/Dockerfile
@@ -1,4 +1,0 @@
-FROM amazoncorretto:17
-WORKDIR /app
-COPY build/libs/*.jar app.jar
-ENTRYPOINT ["java", "-jar", "app.jar"]
