--- auth/build.gradle
@@ -1,44 +1,0 @@
-dependencies {
-    implementation project(':core')
-
-    // Web
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-
-    // DB
-    implementation 'org.springframework.boot:spring-boot-starter-data-jdbc'
-    runtimeOnly 'com.mysql:mysql-connector-j'
-
-    // JPA
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-
-    // JWT
-    implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
-    runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
-    runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
-
-    // Jakarta Bean Validation
-    implementation 'jakarta.validation:jakarta.validation-api:3.0.2'
-    implementation 'org.hibernate.validator:hibernate-validator:8.0.1.Final'
-
-    // Micrometer Tracing
-    implementation 'io.micrometer:micrometer-tracing'
-    implementation 'io.micrometer:micrometer-tracing-bridge-brave'
-
-    // AOP
-    implementation 'org.springframework.boot:spring-boot-starter-aop'
-
-    // Actuator
-    implementation 'org.springframework.boot:spring-boot-starter-actuator'
-
-    // Prometheus
-    implementation 'io.micrometer:micrometer-registry-prometheus'
-}
-
-tasks.bootJar {
-    enabled = true
-}
-
-tasks.jar {
-    enabled = true
-}
