--- auth/src/main/java/kernel360/ckt/auth/application/port/RefreshTokenRepository.java
@@ -1,9 +1,0 @@
-package kernel360.ckt.auth.application.port;
-
-import kernel360.ckt.core.domain.entity.RefreshTokenEntity;
-
-import java.util.Optional;
-
-public interface RefreshTokenRepository {
-    Optional<RefreshTokenEntity> findByCompanyId(Long companyId);
-}
