--- auth/src/main/java/kernel360/ckt/auth/config/CachingRequestBodyFilter.java
@@ -1,22 +1,0 @@
-package kernel360.ckt.auth.config;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-import org.springframework.web.util.ContentCachingRequestWrapper;
-
-import java.io.IOException;
-
-@Component
-public class CachingRequestBodyFilter extends OncePerRequestFilter {
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
-        throws ServletException, IOException {
-        ContentCachingRequestWrapper wrappedRequest = new ContentCachingRequestWrapper(request);
-        filterChain.doFilter(wrappedRequest, response);
-    }
-}
