--- auth/src/main/java/kernel360/ckt/auth/infra/CompanyReadJpaRepository.java
@@ -1,10 +1,0 @@
-package kernel360.ckt.auth.infra;
-
-import kernel360.ckt.core.domain.entity.CompanyEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface CompanyReadJpaRepository extends JpaRepository<CompanyEntity, Long> {
-    Optional<CompanyEntity> findByEmail(String email);
-}
