--- auth/src/main/java/kernel360/ckt/auth/ui/dto/response/TokenResponse.java
@@ -1,6 +1,0 @@
-package kernel360.ckt.auth.ui.dto.response;
-
-public record TokenResponse (
-    String accessToken,
-    String refreshToken
-) {}
