--- auth/src/main/resources/application-local.yml
@@ -1,9 +1,0 @@
-spring:
-  config:
-    activate:
-      on-profile: local
-
-  jpa:
-    hibernate:
-      ddl-auto: update
-    show-sql: true
