--- auth/src/main/resources/application-prod.yml
@@ -1,9 +1,0 @@
-spring:
-  config:
-    activate:
-      on-profile: prod
-
-  jpa:
-    hibernate:
-      ddl-auto: none
-    show-sql: false
