--- auth/src/main/resources/application.yml
@@ -1,41 +1,0 @@
-server:
-  port: 8082
-
-spring:
-  application:
-    name: auth
-
-  datasource:
-    url: jdbc:mysql://${DB_URL}:${DB_PORT}/${DB_DATABASE}
-    username: ${DB_USERNAME}
-    password: ${DB_PASSWORD}
-    driver-class-name: com.mysql.cj.jdbc.Driver
-
-  jpa:
-    show-sql: false
-    open-in-view: false
-    properties.hibernate:
-      naming.physical-strategy: org.hibernate.boot.model.naming.PhysicalNamingStrategySnakeCaseImpl
-      default_batch_fetch_size: 1000
-      globally_quoted_identifiers: false
-
-jwt:
-  secret: ${JWT_SECRET}
-  access-token-expiration: 3600000         # 1시간 (ms)
-  refresh-token-expiration: 604800000      # 7일 (ms)
-
-logging:
-  level:
-    root: INFO
-  pattern:
-    console: "%d{yyyy-MM-dd HH:mm:ss} [%thread] %-5level %logger{36} - %msg%n"
-
-management:
-  endpoints:
-    web:
-      exposure:
-        include: health, info, prometheus
-  metrics:
-    export:
-      prometheus:
-        enabled: true
