--- auth/src/main/resources/logback-spring.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-
-    <include resource="logback.xml" />
-
-    <root level="INFO">
-        <appender-ref ref="APP_FILE"/>
-        <appender-ref ref="ERROR_FILE"/>
-    </root>
-
-</configuration>
