+++ build.gradle
@@ -20,11 +20,13 @@ subprojects {
 
     ext {
         set('springCloudVersion', '2023.0.3')
+        set('springCloudAwsVersion', '3.1.1')
     }
 
     dependencyManagement {
         imports {
             mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
+            mavenBom "io.awspring.cloud:spring-cloud-aws-dependencies:${springCloudAwsVersion}"
         }
     }
