+++ docker-compose.yml
@@ -27,6 +27,7 @@ services:
       SPRING_DATASOURCE_URL: jdbc:mysql://${DB_URL}:3306/${DB_DATABASE}
       SPRING_DATASOURCE_USERNAME: ${DB_USERNAME}
       SPRING_DATASOURCE_PASSWORD: ${DB_PASSWORD}
+      JWT_SECRET: ${JWT_SECRET}
     ports:
       - "8081:8081"
     volumes:
@@ -45,25 +46,6 @@ services:
       JWT_SECRET: ${JWT_SECRET}
     depends_on:
       - admin
-    volumes:
-      - ./logs:/logs
-    networks:
-      - ckt_network
-
-  auth:
-    build:
-      context: ./auth
-    depends_on:
-      - db
-    environment:
-      SPRING_PROFILES_ACTIVE: ${SPRING_PROFILES_ACTIVE}
-      SPRING_REACTOR_CONTEXT_PROPAGATION: AUTO
-      SPRING_DATASOURCE_URL: jdbc:mysql://${DB_URL}:3306/${DB_DATABASE}
-      SPRING_DATASOURCE_USERNAME: ${DB_USERNAME}
-      SPRING_DATASOURCE_PASSWORD: ${DB_PASSWORD}
-      JWT_SECRET: ${JWT_SECRET}
-    ports:
-      - "8082:8082"
     volumes:
       - ./logs:/logs
     networks:
