+++ prometheus/prometheus.yml
@@ -12,11 +12,6 @@ scrape_configs:
     static_configs:
       - targets: ['gateway:8080']
 
-  - job_name: 'auth'
-    metrics_path: '/actuator/prometheus'
-    static_configs:
-      - targets: ['auth:8082']
-
   - job_name: 'collector'
     metrics_path: '/actuator/prometheus'
     static_configs:
