+++ settings.gradle
@@ -3,5 +3,4 @@ rootProject.name = 'ckt-be'
 include 'core'
 include 'admin'
 include 'gateway'
-include 'auth'
 include 'collector'
